requirements.txt
kafka-python
apache-flink
pandas
scikit-learn
flask
streamlit
nltk
python-dotenv